Move ensureNever to a common test trait
